+++ packages/apps/dev-wallet/src/pages/transaction/components/TxContainer.tsx
@@ -40,7 +40,7 @@ export const TxContainer = React.memo(
     );
     const contTx = useSubscribe<ITransaction>(
       'transaction',
-      transaction.continuation?.continuationTxId,
+      localTransaction?.continuation?.continuationTxId,
     );
 
     useEffect(() => {
