+++ packages/apps/dev-wallet/src/utils/normalizeSigs.ts
